+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.0 (27-06-2017)
 - Added `idioms` to the `appicon` configuration
+- Renamed `build-config.yml` to `.natrium.yml`
 
 ## v2.3 (20-03-2017)
 - Add the [Natrium] build phase automatically